use integer for service ranking

+++ osgi-mock/src/main/java/io/wcm/testing/mock/osgi/MockServiceReference.java
@@ -83,8 +83,8 @@ class MockServiceReference implements ServiceReference {
       return 0;
     }
     // sort by decreasing by service ranking, and secondary increasing by service id
-    Long serviceRanking = getServiceRanking();
-    Long otherServiceRanking = ((MockServiceReference)obj).getServiceRanking();
+    Integer serviceRanking = getServiceRanking();
+    Integer otherServiceRanking = ((MockServiceReference)obj).getServiceRanking();
     int serviceRankingCompare = otherServiceRanking.compareTo(serviceRanking);
     if (serviceRankingCompare == 0) {
       Long serviceId = getServiceId();
@@ -106,13 +106,13 @@ class MockServiceReference implements ServiceReference {
     }
   }
 
-  long getServiceRanking() {
+  int getServiceRanking() {
     Number serviceRanking = (Number)getProperty(Constants.SERVICE_RANKING);
     if (serviceRanking != null) {
-      return serviceRanking.longValue();
+      return serviceRanking.intValue();
     }
     else {
-      return 0L;
+      return 0;
     }
   }
